consumer_app/requirements.txt
kafka-python
elasticsearch
configparser
loguru